requirements.txt
fastapi
uvicorn
python-arango
pydantic
jinja2
python-multipart